Expand GF to 2^16

+++ src/gf.h
@@ -4,11 +4,11 @@
 #include <stdint.h>
 
 void gf_init();
-uint8_t gf_add(uint8_t a, uint8_t b);
-uint8_t gf_multiply(uint8_t a, uint8_t b);
-uint8_t gf_divide(uint8_t a, uint8_t b);
-uint8_t gf_inverse(uint8_t a);
-uint8_t gf_generator(int i);
+uint16_t gf_add(uint16_t a, uint16_t b);
+uint16_t gf_multiply(uint16_t a, uint16_t b);
+uint16_t gf_divide(uint16_t a, uint16_t b);
+uint16_t gf_inverse(uint16_t a);
+uint16_t gf_generator(int i);
 
 
 #endif
