+++ src/poly.h
@@ -4,12 +4,12 @@
 #include <stdint.h>
 //highest degree first! a^x .. a0
 
-uint8_t* poly_add(uint8_t* a, int a_len, uint8_t* b, int b_len, int* out_len);
-uint8_t* poly_multiply(uint8_t* a, int a_len, uint8_t* b, int b_len, int* out_len);
-uint8_t poly_evaluate(uint8_t* poly , int len, uint8_t x );
-uint8_t* poly_derivate(uint8_t* poly, int len , int* out_len);
-uint8_t* poly_divide_remainder(uint8_t* a, int a_len, uint8_t* b, int b_len, int* out_len);
-uint8_t* poly_trim(uint8_t* poly, int len, int* out_len);
+uint16_t* poly_add(uint16_t* a, int a_len, uint16_t* b, int b_len, int* out_len);
+uint16_t* poly_multiply(uint16_t* a, int a_len, uint16_t* b, int b_len, int* out_len);
+uint16_t poly_evaluate(uint16_t* poly , int len, uint16_t x );
+uint16_t* poly_derivate(uint16_t* poly, int len , int* out_len);
+uint16_t* poly_divide_remainder(uint16_t* a, int a_len, uint16_t* b, int b_len, int* out_len);
+uint16_t* poly_trim(uint16_t* poly, int len, int* out_len);
